Add dynamic age calculation to About Me section
In index.js, a function was added to calculate the current age based on the birth date. This ensures that age is always correct and automatically updated without needing manual edits. In index.html, '22-year-old' string was replaced by an dynamic HTML span element, which will be populated by the JS function. This change is made to improve the maintenance and readability of the site.