Update code format and improve logo color transition
Code on various web pages was reformatted for easier readability, this involved adding line breaks and indentation to longer sections. Additionally, the color transition on the '.dev' text part of the website logo was improved. Instead of a blinking effect, colors will now increment in hue gradually, creating a smoother transition. This achieves a more desirable aesthetic for the logo and enhances user experience. Regular intervals were also set for the incrementation.