Add mobile responsiveness to CSS
Added media queries to the CSS in order to improve the styling of the website on mobile devices. This will create a better user experience on smaller screens, as elements will now be properly aligned and displayed. Aimed particularly at devices with max width of 600px. Changed aspects include body padding, header and social icon layout, and content justification in skills and projects sections.