Update PHP configuration and modify css styling
Updated the PHP language level to 8.2 in the project configuration. This is to make sure we are utilizing the new and updated features provided by PHP 8.2.

Changed the padding and dimensions in style.css. This was done to allow better fit and positioning of elements, enhancing the layout and responsiveness for various devices. Modified the max-width and max-height of the ".project-bubble-img", ".skill-bubble-img", and ".php-skill-img" classes and revised the padding of several components to improve visual appearance.